Upgrade deps and enable npm
This PR is intended to make the web compatible with npm-only usage.

Nowadays npm is as full featured as yarn and most people tend to use only npm for development.

npm has a stronger dependency check, it was complaining of some unmet dependency requirements, so this PR also upgrades the deps and include the package-lock.json file for npm.